changed release year datatype from double to int

+++ app/src/main/java/com/werelit/neurolls/neurolls/model/Book.java
@@ -11,7 +11,7 @@ public class Book extends Media{
     /** description of the book */
     private String description;
 
-    public Book(String mediaName, String mediaGenre, double releaseYear,
+    public Book(String mediaName, String mediaGenre, int releaseYear,
                 String author, String publisher, String description) {
 
         super(mediaName, mediaGenre, releaseYear);
